Remove deps on action_dispatch and activesupport
This gem has fairly strict dependency requirements, as
it is used in the turbot client that we send out to bot
authors. This has a tightly specified set of dependencies
to reflect the environment it runs under on the server.
Taking away the dependency on activesupport allows us to
work with the openc_bot gem, which in turn is tied to the main
app which runs activesupport 3.2.17